updation in delete project api

+++ brightStafferapp/views.py
@@ -358,6 +358,7 @@ class TopProjectList(generics.ListCreateAPIView):
         del (response.data['results'])
         return response
 
+
 class ProjectDelete(generics.ListCreateAPIView):
     queryset = Projects.objects.all()
     serializer_class = ProjectSerializer
@@ -367,14 +368,17 @@ class ProjectDelete(generics.ListCreateAPIView):
         queryset = super(ProjectDelete, self).get_queryset()
         recruiter = self.request.query_params.get('recruiter')
         token = self.request.query_params.get('token')
+        values = Token.objects.filter(user__username=recruiter, key=token)
+        if not values:
+            return util.returnErrorShorcut(403, 'Either Recruiter Email or Token id is not valid')
         project_id_list = self.request.query_params.get('project').split(',')  # ('talent[]')[0].split(',')
-
         for project_id in project_id_list:
             talent_objs = Projects.objects.filter(id=project_id)
             if not talent_objs:
                 return util.returnErrorShorcut(403, 'Project with id {} dosen\'t exist in database.'.format(project_id))
             deleted = Projects.objects.filter(id=project_id).delete()
-        return queryset
+            if deleted:
+                return queryset.filter(recruiter__username=recruiter)
 
 
 class UpdateRecruiter(View):
